+++ tests/SimpleThings/Tests/EntityAudit/Fixtures/Relation/BookCategory.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace SimpleThings\EntityAudit\Tests\Fixtures\Relation;
+
+use Doctrine\ORM\Mapping as ORM;
+
+/**
+ * @ORM\Entity
+ */
+class BookCategory extends Category {}
