+++ tests/SimpleThings/Tests/EntityAudit/Fixtures/Relation/OneToOneAuditedEntity.php
@@ -1,0 +1,43 @@
+<?php
+
+namespace SimpleThings\EntityAudit\Tests\Fixtures\Relation;
+
+use Doctrine\ORM\Mapping as ORM;
+
+/** @ORM\Entity */
+class OneToOneAuditedEntity
+{
+    /** @ORM\Id @ORM\Column(type="integer") @ORM\GeneratedValue(strategy="AUTO") */
+    protected $id;
+
+    /** @ORM\Column(type="string") */
+    protected $title;
+
+    /** @ORM\OneToOne(targetEntity="OneToOneMasterEntity") */
+    protected $master;
+
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    public function getTitle()
+    {
+        return $this->title;
+    }
+
+    public function setTitle($title)
+    {
+        $this->title = $title;
+    }
+
+    public function getMaster()
+    {
+        return $this->master;
+    }
+
+    public function setMaster($master)
+    {
+        $this->master = $master;
+    }
+}
