+++ tests/SimpleThings/Tests/EntityAudit/Fixtures/Relation/SomeEntity.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace SimpleThings\EntityAudit\Tests\Fixtures\Relation;
+
+use Doctrine\ORM\Mapping as ORM;
+
+/**
+ * @ORM\MappedSuperclass()
+ */
+abstract class SomeEntity
+{
+    /** @ORM\Id @ORM\Column(type="integer") @ORM\GeneratedValue */
+    private $id;
+
+    public function getId()
+    {
+        return $this->id;
+    }
+}
