+++ tests/SimpleThings/Tests/EntityAudit/Fixtures/Relation/UnManagedIndexByOwner.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace SimpleThings\EntityAudit\Tests\Fixtures\Relation;
+
+class UnManagedIndexByOwner
+{
+
+}
